sql/requirements.txt
Faker==40.1.2
numpy==2.4.1
pandas==3.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
